requirements.txt
# Python >=3.10, <3.14 required
crewai[tools]>=1.13.0
langchain-groq>=0.3.0
fastapi>=0.111.0
uvicorn>=0.29.0
streamlit>=1.35.0
python-dotenv>=1.0.1
requests>=2.32.2
duckduckgo-search>=6.0.0
pydantic>=2.7.1
httpx>=0.27.0
